Renamed number insight files to start with ni-, + split advanced async into two snippets
