+++ number-insight/ni-basic.go
