+++ number-insight/ni-standard.go
